Prefix Optimiser classes with Optimiser and next Integer and Boolean variables
