+++ traversal/optimiser/OptimiserConstraint.java
@@ -24,18 +24,18 @@ import com.google.ortools.linearsolver.MPSolver;
 import java.util.HashMap;
 import java.util.Map;
 
-public class Constraint {
+public class OptimiserConstraint {
 
     private final double lowerBound;
     private final double upperBound;
     private final String name;
-    private final Map<Variable, Double> coefficients;
+    private final Map<OptimiserVariable<?>, Double> coefficients;
     private MPConstraint mpConstraint;
     private State state;
 
-    private enum State { INACTIVE, ACTIVE }
+    private enum State {INACTIVE, ACTIVE}
 
-    public Constraint(double lowerBound, double upperBound, String name) {
+    public OptimiserConstraint(double lowerBound, double upperBound, String name) {
         this.lowerBound = lowerBound;
         this.upperBound = upperBound;
         this.coefficients = new HashMap<>();
@@ -43,7 +43,7 @@ public class Constraint {
         this.state = State.INACTIVE;
     }
 
-    public void setCoefficient(Variable variable, double coeff) {
+    public void setCoefficient(OptimiserVariable<?> variable, double coeff) {
         assert state == State.INACTIVE;
         coefficients.put(variable, coeff);
     }
