--- traversal/optimiser/Variable.java
@@ -1,49 +1,0 @@
-/*
- * Copyright (C) 2021 Vaticle
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-package com.vaticle.typedb.core.traversal.optimiser;
-
-import com.google.ortools.linearsolver.MPSolver;
-import com.google.ortools.linearsolver.MPVariable;
-
-public abstract class Variable<T> {
-
-    enum State { INACTIVE, ACTIVE }
-
-    final String name;
-
-    Variable(String name) {
-        this.name = name;
-    }
-
-    public abstract T solutionValue();
-
-    abstract MPVariable mpVariable();
-
-    abstract void recordValue();
-
-    public abstract void clearInitial();
-
-    abstract boolean hasInitial();
-
-    public abstract double getInitial();
-
-    abstract void activate(MPSolver solver);
-
-    abstract void deactivate();
-}
